gcc_mt_m11_v3/requirements.txt
# Core dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-dotenv==1.0.0
pytest==7.4.4

# Logging and utilities
structlog==24.1.0

# Database - PostgreSQL (Strategy 1: RLS)
psycopg2-binary==2.9.9

# Vector Database - Pinecone (Strategy 2: Namespace Isolation)
pinecone-client==3.0.0

# Cache - Redis (Key Prefix Isolation)
redis==5.0.1

# AWS Services (S3 Prefix Isolation)
boto3==1.34.0

# Jupyter for notebooks
jupyter==1.0.0
ipykernel==6.29.0
notebook==7.0.0

# Testing
pytest-asyncio==0.23.3
